pyproject.toml
[tool.poetry]
name = "zindi-soil-hyperspectral"
version = "0.1.0"
description = ""
authors = ["Daniel"]
readme = "README.md"
packages = [{include = "zindi_soil_hyperspectral"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
numpy = "^1.26.1"
pandas = "^2.1.2"
matplotlib = "^3.8.1"
seaborn = "^0.13.0"
ipykernel = "^6.26.0"
ngboost = "^0.4.2"
cmaes = "^0.10.0"
pymoo = "^0.6.0.1"
optuna = "^3.4.0"
plotly = "^5.18.0"
umap-learn = "^0.5.4"
opentsne = "^1.0.0"
kaleido = "0.2.1"
mbtr = "^0.1.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
